pyproject.toml
[project]
name = "anthropic-getting-started"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.49.0",
    "python-dotenv>=1.0.1",
]
